[a16wmix] Move host launcher/tile-config/CSV out of the kernel package into tests
The kernel package should hold kernels only. flydsl_a16w4_gemm{1,2}, the tile
resolvers, CSV parsing, and a16wi4 host helpers were the entire content of
kernels/moe/moe_2stage_a16wmix/__init__.py and were imported only by the test.
Move them verbatim to tests/kernels/moe_a16wmix_host.py (relative .gemm1/.gemm2
imports -> absolute); the package __init__.py is now a minimal marker. Kernel
builders (compile_gemm1/2_a16w4_port) + shared helpers (utils.py) stay in the
package. test_moe_gemm.py imports the host helpers from the new location.

Behavior-preserving: a16w4 script + pytest unchanged (9 passed, 6 xfailed).
