+++ kernels/moe/moe_2stage_a16wmix/gemm2.py
@@ -10,7 +10,7 @@ from flydsl.expr.typing import T
 from flydsl.expr.typing import Vector as Vec
 from kernels.common.layout_utils import crd2idx
 
-from .gemm1 import (
+from .utils import (
     A16WI4_GROUP_SIZE,
     _a16w4_swizzle_xor16,
     _buffer_i32_scalar_read,
